requirements.txt
faster_whisper==1.0.3
Flask==3.0.3
Flask_Cors==4.0.1
ollama==0.2.1
psycopg2_binary==2.9.9
python-dotenv==1.0.1
pytubefix==6.1.1
